# [SPARK-35384][SQL] Improve performance for InvokeLike.invoke