--- a/sql/core/benchmarks/V2FunctionBenchmark-results.txt
+++ b/sql/core/benchmarks/V2FunctionBenchmark-results.txt
@@ -2,43 +2,43 @@ OpenJDK 64-Bit Server VM 1.8.0_292-b10 on Linux 5.4.0-1046-azure
 Intel(R) Xeon(R) CPU E5-2673 v3 @ 2.40GHz
 scalar function (long + long) -> long, result_nullable = true codegen = true:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ------------------------------------------------------------------------------------------------------------------------------------------------------------
-native_long_add                                                                       11275          12171         998         44.3          22.5       1.0X
-java_long_add_default                                                                 37040          37728         755         13.5          74.1       0.3X
-java_long_add_magic                                                                   13974          14218         276         35.8          27.9       0.8X
-java_long_add_static_magic                                                            12019          12260         248         41.6          24.0       0.9X
-scala_long_add_default                                                                38262          38520         228         13.1          76.5       0.3X
-scala_long_add_magic                                                                  13940          14062         202         35.9          27.9       0.8X
+native_long_add                                                                        9474          10354         975         52.8          18.9       1.0X
+java_long_add_default                                                                 36465          36994         607         13.7          72.9       0.3X
+java_long_add_magic                                                                   11922          12228         433         41.9          23.8       0.8X
+java_long_add_static_magic                                                            10596          10867         275         47.2          21.2       0.9X
+scala_long_add_default                                                                36644          37225         705         13.6          73.3       0.3X
+scala_long_add_magic                                                                  11887          12140         222         42.1          23.8       0.8X
 
 OpenJDK 64-Bit Server VM 1.8.0_292-b10 on Linux 5.4.0-1046-azure
 Intel(R) Xeon(R) CPU E5-2673 v3 @ 2.40GHz
 scalar function (long + long) -> long, result_nullable = false codegen = true:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 -------------------------------------------------------------------------------------------------------------------------------------------------------------
-native_long_add                                                                        11041          11194         133         45.3          22.1       1.0X
-java_long_add_default                                                                  35213          35716         454         14.2          70.4       0.3X
-java_long_add_magic                                                                    13857          13945          78         36.1          27.7       0.8X
-java_long_add_static_magic                                                             11272          11382          97         44.4          22.5       1.0X
-scala_long_add_default                                                                 35157          36078        1103         14.2          70.3       0.3X
-scala_long_add_magic                                                                   14005          14194         223         35.7          28.0       0.8X
+native_long_add                                                                         9197           9295         121         54.4          18.4       1.0X
+java_long_add_default                                                                  32693          32752          76         15.3          65.4       0.3X
+java_long_add_magic                                                                    11981          12463         711         41.7          24.0       0.8X
+java_long_add_static_magic                                                              9447           9529          87         52.9          18.9       1.0X
+scala_long_add_default                                                                 31355          31600         220         15.9          62.7       0.3X
+scala_long_add_magic                                                                   11672          11834         159         42.8          23.3       0.8X
 
 OpenJDK 64-Bit Server VM 1.8.0_292-b10 on Linux 5.4.0-1046-azure
 Intel(R) Xeon(R) CPU E5-2673 v3 @ 2.40GHz
 scalar function (long + long) -> long, result_nullable = true codegen = false:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 -------------------------------------------------------------------------------------------------------------------------------------------------------------
-native_long_add                                                                        33550          33648         144         14.9          67.1       1.0X
-java_long_add_default                                                                  44094          44832         690         11.3          88.2       0.8X
-java_long_add_magic                                                                   173631         175419        1611          2.9         347.3       0.2X
-java_long_add_static_magic                                                            173095         174496        1530          2.9         346.2       0.2X
-scala_long_add_default                                                                 46445          46774         430         10.8          92.9       0.7X
-scala_long_add_magic                                                                  174399         175016         550          2.9         348.8       0.2X
+native_long_add                                                                        31047          31422         362         16.1          62.1       1.0X
+java_long_add_default                                                                  40283          40423         121         12.4          80.6       0.8X
+java_long_add_magic                                                                    54133          54210          67          9.2         108.3       0.6X
+java_long_add_static_magic                                                             56368          56608         212          8.9         112.7       0.6X
+scala_long_add_default                                                                 40166          40296         132         12.4          80.3       0.8X
+scala_long_add_magic                                                                   55704          55902         249          9.0         111.4       0.6X
 
 OpenJDK 64-Bit Server VM 1.8.0_292-b10 on Linux 5.4.0-1046-azure
 Intel(R) Xeon(R) CPU E5-2673 v3 @ 2.40GHz
 scalar function (long + long) -> long, result_nullable = false codegen = false:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 --------------------------------------------------------------------------------------------------------------------------------------------------------------
-native_long_add                                                                         36506          36656         251         13.7          73.0       1.0X
-java_long_add_default                                                                   47151          47540         370         10.6          94.3       0.8X
-java_long_add_magic                                                                    178691         182457        1327          2.8         357.4       0.2X
-java_long_add_static_magic                                                             177151         178258        1151          2.8         354.3       0.2X
-scala_long_add_default                                                                  47127          47760         666         10.6          94.3       0.8X
-scala_long_add_magic                                                                   179941         181417        1545          2.8         359.9       0.2X
+native_long_add                                                                         29897          30342         568         16.7          59.8       1.0X
+java_long_add_default                                                                   40628          41075         664         12.3          81.3       0.7X
+java_long_add_magic                                                                     54553          54755         182          9.2         109.1       0.5X
+java_long_add_static_magic                                                              55410          55532         127          9.0         110.8       0.5X
+scala_long_add_default                                                                  39775          40392         661         12.6          79.5       0.8X
+scala_long_add_magic                                                                    54221          54631         360          9.2         108.4       0.6X